FORTUNE DISCUSSES THE DIFFERENCES BETWEEN MEN AND WOMEN:	#6
Bathrooms:
	A man has six items in his bathroom -- a toothbrush, toothpaste,
shaving cream, razor, a bar of Dial soap, and a towel from the Holiday Inn.
The average number of items in the typical woman's bathroom is 437.  A man
would not be able to identify most of these items.

Groceries:
	A woman makes a list of things she needs and then goes to the store
and buys these things.  A man waits 'til the only items left in his fridge
are half a lime and a Blue Ribbon.  Then he goes grocery shopping.  He buys
everything that looks good.  By the time a man reaches the checkout counter,
his cart is packed tighter that the Clampett's car on Beverly Hillbillies.
Of course, this will not stop him from entering the 10-items-or-less lane.